+++ Changes.md
@@ -9,6 +9,7 @@ Version 4.0.0.1 - TBR
 5. Fix 'Course overview' dropdown menu item 'All (except removed from view)' icon.
 6. Update FontAwesome from 5.12.2 to 5.12.4.
 7. Add FontAwesome 6.0.0-beta.
+8. Add Morph, Quartz, Vapor and Zephyr bootswatches.
 
 Version 3.11.1.4 - 08/10/21
 ---------------------------